refactor and embellish server.cpp
